requirements.txt
flask
scikit-learn
joblib
beautifulsoup4
requests
numpy
pandas
python-whois
dnspython
